Move BaseSttpStacClient into shared folder
